daemon/requirements.txt
PyYAML==6.0
sentry-usage-accountant==0.0.11
